Refactor into a collector base class + dedicated prometheus and move cli to main cli with arguments
